# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMART CLINIC - Healthcare SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    post:
      operationId: createClinic
      tags: [clinics]
      summary: Create a new clinic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicInput"
      responses:
        "201":
          description: Clinic created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/lookup:
    get:
      operationId: lookupClinic
      tags: [clinics]
      summary: Look up clinic by code
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clinic found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicPublic"
        "404":
          description: Clinic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/mine:
    get:
      operationId: getMyClinic
      tags: [clinics]
      summary: Get clinic for current user
      responses:
        "200":
          description: Clinic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not in a clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clinics/{clinicId}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get clinic by ID
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clinic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
    patch:
      operationId: updateClinic
      tags: [clinics]
      summary: Update clinic settings
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicUpdate"
      responses:
        "200":
          description: Updated clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"

  # ===== JOIN REQUESTS / MEMBERS =====
  /clinics/{clinicId}/join-requests:
    post:
      operationId: createJoinRequest
      tags: [members]
      summary: Request to join a clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequestInput"
      responses:
        "201":
          description: Join request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
    get:
      operationId: listJoinRequests
      tags: [members]
      summary: List pending join requests for clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of join requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
  /clinics/{clinicId}/join-requests/{requestId}:
    patch:
      operationId: updateJoinRequest
      tags: [members]
      summary: Approve or reject a join request
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequestUpdate"
      responses:
        "200":
          description: Updated join request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
  /clinics/{clinicId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List clinic members
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicMember"
  /clinics/{clinicId}/members/{userId}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update clinic member (role, status)
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicMember"
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member from clinic
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ===== PATIENTS =====
...
  /clinics/{clinicId}/doctors:
    get:
      operationId: listClinicDoctors
      tags: [clinics]
      summary: List doctors in a clinic for appointment booking
      parameters:
        - in: path
          name: clinicId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of clinic doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicDoctor"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ClinicInput:
      type: object
      required: [name, type, address, city, province, contactNumber, email]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [general, specialist, dental, pediatric, mental_health, other]
        clinicType:
          type: string
          enum: [private, government, ngo]
        address:
          type: string
        city:
          type: string
        province:
          type: string
        contactNumber:
          type: string
        email:
          type: string
          format: email
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    ClinicUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        clinicType:
          type: string
          enum: [private, government, ngo]
        address:
          type: string
        city:
          type: string
        province:
          type: string
        contactNumber:
          type: string
        email:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isActive:
          type: boolean
        billingEnabled:
          type: boolean
        consultationFeeEnabled:
          type: boolean
        pharmacyBillingEnabled:
          type: boolean
        labBillingEnabled:
          type: boolean
    Clinic:
      type: object
      required: [id, name, code, type, clinicType, address, city, province, contactNumber, email, isActive, billingEnabled, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        type:
          type: string
        clinicType:
          type: string
          enum: [private, government, ngo]
        address:
          type: string
        city:
          type: string
        province:
          type: string
        contactNumber:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        billingEnabled:
          type: boolean
        consultationFeeEnabled:
          type: boolean
        pharmacyBillingEnabled:
          type: boolean
        labBillingEnabled:
          type: boolean
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        whatsappEnabled:
          type: boolean
          description: Whether WhatsApp notifications are enabled (credentials are never returned on this schema)
        whatsappProvider:
          type: string
          enum: [meta, twilio, 360dialog]
...
    ClinicPublic:
      type: object
      required: [id, name, code, type, city]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        type:
          type: string
        city:
          type: string
        province:
          type: string
...
    JoinRequestInput:
      type: object
      required: [requestedRole]
      properties:
        requestedRole:
          type: string
          enum: [doctor, nurse, receptionist, pharmacist, lab_technician, cashier]
        message:
          type: ["string", "null"]
    JoinRequestUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        assignedRole:
          type: ["string", "null"]
    JoinRequest:
      type: object
      required: [id, clinicId, userId, requestedRole, status, createdAt]
      properties:
        id:
          type: string
        clinicId:
          type: string
        userId:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        requestedRole:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    MemberUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [clinic_admin, doctor, nurse, receptionist, pharmacist, lab_technician, cashier]
        status:
          type: string
          enum: [active, inactive, suspended]
        reason:
          type: string
          description: Optional reason for status change (used in audit log)

    # ===== PATIENT SCHEMAS =====
...